+++ vscode/webviews/components/shadcn/ui/tabs.tsx
@@ -12,5 +12,11 @@ export const TabContainer = React.forwardRef<
     React.ElementRef<typeof Tabs.Content>,
     React.ComponentPropsWithoutRef<typeof Tabs.Content>
 >(({ className, ...props }, ref) => {
-    return <Tabs.Content ref={ref} {...props} className="tw-h-full tw-flex tw-flex-col tw-gap-4" />
+    return (
+        <Tabs.Content
+            ref={ref}
+            {...props}
+            className="tw-h-full tw-flex tw-flex-col tw-overflow-auto tw-gap-4"
+        />
+    )
 })